requirements.txt
streamlit==1.23.1
openai
langchain
gradio
python-dotenv
pymysql
pymilvus
tiktoken
loguru